Add 'postgres' flag and dummy Postgres module
Add 'postgres' flag to:
1. `make` targets:
  1. `make configure`
  1. `make /dist/bin/kupo`
1. kupo.cabal:
  1. Conditionally include the Postgres module vs the SQLite module
  1. Add `cpp-options: -Dpostgres`
1. Kupo.App.Database: import Postgres or SQLite module depending on flag

Also add additional files to `make clean` to cause the project to be recompiled after `clean`
