+++ kupo.cabal
@@ -42,6 +42,11 @@ flag production
   manual: True
   default: False
 
+flag postgres
+  description: Compile for use with a PostgresSQL DB instead of sqlite
+  manual: True
+  default: False
+
 library
   exposed-modules:
       Kupo
@@ -116,8 +121,15 @@ library
       Kupo.Prelude
       Kupo.Version
       Kupo.Version.TH
-  other-modules:
+  if flag(postgres)
+    other-modules:
       Paths_kupo
+      Kupo.App.Database.Postgres
+  else
+    other-modules:
+      Paths_kupo
+      Kupo.App.Database.SQLite
+
   hs-source-dirs:
       src
   default-extensions:
@@ -228,6 +240,9 @@ library
   default-language: Haskell2010
   if flag(production)
     ghc-options: -Wunused-packages -Werror -O2
+  if flag(postgres)
+    cpp-options: -Dpostgres
+
 
 executable kupo
   main-is: Main.hs
@@ -279,6 +294,8 @@ executable kupo
   default-language: Haskell2010
   if flag(production)
     ghc-options: -O2 -Wunused-packages -Werror "-with-rtsopts=-N -A16m -qb -qg"
+  if flag(postgres)
+    cpp-options: -Dpostgres
 
 test-suite unit
   type: exitcode-stdio-1.0
